requirements.txt
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-Cors==3.0.10
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
PyJWT==2.4.0
pymongo==4.2.0
python-dotenv==0.20.0
six==1.16.0
Werkzeug==2.2.2
